+++ bin/node/src/commands/bundled.rs
@@ -70,7 +70,7 @@ pub enum BundledCommand {
         ///
         /// This can be further configured using environment variables as defined in the official
         /// OpenTelemetry documentation. See our operator manual for further details.
-        #[arg(long = "enable-otel", default_value_t = false, env = ENV_ENABLE_OTEL, value_name = "bool")]
+        #[arg(long = "enable-otel", default_value_t = false, env = ENV_ENABLE_OTEL, value_name = "BOOL")]
         open_telemetry: bool,
 
         /// Interval at which to produce blocks in milliseconds.