+++ bin/node/src/commands/store.rs
@@ -25,6 +25,9 @@ use super::{
 };
 use crate::system_monitor::SystemMonitor;
 
+/// The default filepath for the genesis account.
+const DEFAULT_ACCOUNT_PATH: &str = "account.mac";
+
 #[derive(clap::Subcommand)]
 pub enum StoreCommand {
     /// Bootstraps the blockchain database with the genesis block.
@@ -56,7 +59,7 @@ pub enum StoreCommand {
         ///
         /// This can be further configured using environment variables as defined in the official
         /// OpenTelemetry documentation. See our operator manual for further details.
-        #[arg(long = "enable-otel", default_value_t = false, env = ENV_ENABLE_OTEL, value_name = "bool")]
+        #[arg(long = "enable-otel", default_value_t = false, env = ENV_ENABLE_OTEL, value_name = "BOOL")]
         open_telemetry: bool,
 
         /// Interval at which to monitor the system in milliseconds.
@@ -128,7 +131,7 @@ impl StoreCommand {
         //
         // Without this the accounts would be inaccessible by the user.
         // This is not used directly by the node, but rather by the owner / operator of the node.
-        let filepath = accounts_directory.join("account.mac");
+        let filepath = accounts_directory.join(DEFAULT_ACCOUNT_PATH);
         File::create_new(&filepath)
             .and_then(|mut file| file.write_all(&account_file.to_bytes()))
             .with_context(|| {